LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-hi-expected.txt

FAIL navigator.language should be hi. Was hi-in.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
